moved banner to utils package
